new function in main for determining best combination of ai. Game.play now can be silent. new class Result for holding and sorting results information. Everything yet to be tested.

+++ src/players/ai_players/support_classes/PoweredLengthCooValEstimator.java
@@ -5,9 +5,9 @@ import grid_computations.PotentialStreak;
 
 public class PoweredLengthCooValEstimator extends AbstractCooValFromStreakEstimator {
 
-	private int power;
+	private double power;
 	
-	public PoweredLengthCooValEstimator(int power) {
+	public PoweredLengthCooValEstimator(double power) {
 		this.power = power;
 	}
